+++ fcxref/find/__init__.py
@@ -1,5 +1,5 @@
 from .make_find import make_find
-from .property import Property
+from .query import Query
 from .reference import Reference
 
-__all__ = ['make_find', 'Property', 'Reference']
+__all__ = ['make_find', 'Query', 'Reference']