+++ fcxref/find/reference.py
@@ -2,9 +2,8 @@ import os
 
 __all__ = ['Reference']
 
-
 class Reference:
-    """Represents a reference to a property in another document."""
+    """Represents a reference to a property or object in another document."""
 
     def __init__(self,
                  document_path: str,
@@ -13,6 +12,7 @@ class Reference:
                  reference_attribute: str,
                  location: str,
                  match: str,
+                 content: str,
                  xpath: str) -> None:
         self.document_path = document_path
         self.object_name = object_name
@@ -20,13 +20,23 @@ class Reference:
         self.reference_attribute = reference_attribute
         self.location = location
         self.match = match
+        self.content = content
         self.xpath = xpath
 
     def __str__(self):
         return self._to_string()
 
     def __repr__(self):
-        return self._to_string()
+        return (
+            f'Reference("{self.document_path}",\n' +
+            f'          "{self.object_name}",\n' +
+            f'          "{self.property_name}",\n' +
+            f'          "{self.reference_attribute}",\n' +
+            f'          "{self.location}",\n' +
+            f'          "{self.match}",\n' +
+            f'          "{self.content}",\n' +
+            f'          "{self.xpath}")'
+        )
 
     def _to_string(self):
         return '{} {}.{} ({})'.format(
@@ -44,6 +54,7 @@ class Reference:
                 self.reference_attribute == o.reference_attribute and
                 self.location == o.location and
                 self.match == o.match and
+                self.content == o.content and
                 self.xpath == o.xpath
             )
         return False