+++ fcxref/rename/make_rename.py
@@ -1,7 +1,7 @@
 from typing import Callable, Dict, List, Tuple
 from xml.etree.ElementTree import Element
 
-from ..find import Property, Reference, make_find
+from ..find import Query, Reference, make_find
 from ..group_references_by_document_path import \
     group_references_by_document_path
 from .rename_owner_document import rename_owner_document
@@ -17,8 +17,8 @@ def make_rename(find_root_by_document_path: Callable[[str], Dict[str, Element]])
                from_to_properties: Tuple[str, str]) -> Dict[str, Element]:
         from_property_name, to_property_name = from_to_properties
         root_by_document_path = find_root_by_document_path(base_path)
-        from_property = Property(document, object_name, from_property_name)
-        to_property = Property(document, object_name, to_property_name)
+        from_property = Query(document, object_name, from_property_name)
+        to_property = Query(document, object_name, to_property_name)
         references = find(base_path, from_property)
         references_by_document_path = group_references_by_document_path(
             references)
@@ -35,7 +35,7 @@ def make_rename(find_root_by_document_path: Callable[[str], Dict[str, Element]])
 
 def rename_references_in_document_xml(root_by_document_path: Dict[str, Element],
                                       references_by_document_path: Dict[str, List[Reference]],
-                                      to_property: Property) -> Dict[str, Element]:
+                                      to_property: Query) -> Dict[str, Element]:
     renamed_root_by_document_path = {}
     for document_path, references in references_by_document_path.items():
         root = root_by_document_path[document_path]