+++ fcxref/rename/rename_owner_document.py
@@ -3,7 +3,7 @@ import re
 from typing import Callable, Dict, Optional, Tuple
 from xml.etree.ElementTree import Element
 
-from ..find import Property
+from ..find import Query
 from ..find.find_references_in_root import find_references_in_root
 from ..rename.rename_references_in_root import rename_references_in_root
 from .label import extract_label, is_label
@@ -13,7 +13,7 @@ logger = logging.getLogger(__name__)
 
 def rename_owner_document(find_root_by_document_path: Callable[[str, str], Dict[str, Element]],
                           base_path: str,
-                          from_property: Property,
+                          from_property: Query,
                           to_property_name: str) -> Dict[str, Element]:
     document = from_property.document
     find_root = find_document_by_label if is_label(
@@ -35,7 +35,7 @@ def rename_owner_document(find_root_by_document_path: Callable[[str, str], Dict[
         cell_element.set('alias', to_property_name)
         pattern = re.compile(r'\b{}\b'.format(from_property.property_name))
         references = find_references_in_root(document_path, root, pattern)
-        to_property = Property(from_property.document,
+        to_property = Query(from_property.document,
                                from_property.object_name,
                                to_property_name)
         copy = rename_references_in_root(root, references, to_property)